# Server/openapi_chatgpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Options Terminal API",
    "description": "Professional options analysis API for quantitative traders. Compute Black-Scholes Greeks, extract risk-neutral distributions, and model complex payoff profiles.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://tenuously-unswitched-alma.ngrok-free.dev"
    }
  ],
  "paths": {
    "/api/mcp/call-tool": {
      "post": {
        "summary": "Execute options analysis tool",
        "description": "Call tool with params. REQUIRED: get_expirations(underlying), get_chain(underlying,expiration), compute_greeks(underlying,expiration), get_distribution(underlying,expiration), compute_payoff_profile(side,option_type,underlying,strike,expiration)",
        "operationId": "callTool",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tool", "arguments"],
                "properties": {
                  "tool": {
                    "type": "string",
                    "enum": [
                      "get_expirations",
                      "get_chain",
                      "compute_greeks",
                      "get_distribution",
                      "compute_payoff_profile",
                      "get_historical_prices_tool"
                    ]
                  },
                  "arguments": {
                    "type": "object",
                    "required": ["underlying"],
                    "properties": {
                      "underlying": {
                        "type": "string",
                        "description": "REQUIRED for ALL tools. Stock ticker (e.g., 'AAPL', 'TSLA', 'NVDA')"
                      },
                      "expiration": {
                        "type": "string",
                        "description": "REQUIRED for get_chain, compute_greeks, get_distribution, compute_payoff_profile. Format: YYYY-MM-DD"
                      },
                      "side": {
                        "type": "string",
                        "enum": ["long", "short"],
                        "description": "REQUIRED for compute_payoff_profile only"
                      },
                      "option_type": {
                        "type": "string",
                        "enum": ["call", "put"],
                        "description": "REQUIRED for compute_payoff_profile only"
                      },
                      "strike": {
                        "type": "number",
                        "description": "REQUIRED for compute_payoff_profile only. Strike price"
                      },
                      "min_moneyness": {
                        "type": "number",
                        "description": "OPTIONAL for get_distribution. Minimum strike/spot ratio (default: 0.7)"
                      },
                      "max_moneyness": {
                        "type": "number",
                        "description": "OPTIONAL for get_distribution. Maximum strike/spot ratio (default: 1.3)"
                      },
                      "spot_min": {
                        "type": "number",
                        "description": "OPTIONAL for compute_payoff_profile. Minimum spot price for payoff range"
                      },
                      "spot_max": {
                        "type": "number",
                        "description": "OPTIONAL for compute_payoff_profile. Maximum spot price for payoff range"
                      },
                      "period": {
                        "type": "string",
                        "description": "OPTIONAL for get_historical_prices_tool. Time period: '1mo', '3mo', '6mo', '1y', etc."
                      },
                      "interval": {
                        "type": "string",
                        "description": "OPTIONAL for get_historical_prices_tool. Interval: '1d', '1h', '1wk', etc."
                      }
                    }
                  }
                }
              },
              "examples": {
                "get_expirations": {
                  "value": {
                    "tool": "get_expirations",
                    "arguments": {
                      "underlying": "TSLA"
                    }
                  }
                },
                "get_chain": {
                  "value": {
                    "tool": "get_chain",
                    "arguments": {
                      "underlying": "AAPL",
                      "expiration": "2025-12-19"
                    }
                  }
                },
                "compute_payoff_profile": {
                  "value": {
                    "tool": "compute_payoff_profile",
                    "arguments": {
                      "side": "long",
                      "option_type": "call",
                      "underlying": "NVDA",
                      "strike": 500.0,
                      "expiration": "2026-01-16"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful tool execution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required arguments"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {}
  }
}